Add coaches template and .md
