+++ src/pages/coaches/alex-lobera.md
@@ -1,0 +1,16 @@
+---
+name: Alex Lobera
+title: Tech Lead at LeanJS
+gitHub: https://github.com/alexlbr
+medium: https://medium.com/@alex_lobera
+twitter: https://twitter.com/alex_Lobera
+linkedIn: https://uk.linkedin.com/in/alexlobera
+meetupGroup: https://www.meetup.com/JavaScript-london/
+aboutBullets:
+  - Full-stack JavaScript Developer passionate about <strong>teaching React since <a href="https://www.meetup.com/JavaScript-London/events/230287691/">May, 2016</a></strong>.
+  - Alex holds a <strong>computer science degree</strong> and he is also a <strong>certified teacher</strong> in Spanish language.
+  - Organizer of the <a href="https://www.meetup.com/JavaScript-london/">JavaScript London Meetup</a> and other popular meetups in the EU.
+blockquote: Teaching other developers is very rewarding. I love sharing what I've learned in my career to help others grow and give back. I think this is what really makes our industry move forward.
+youtubeVideoId: QiR8iNq3tCQ
+videoDescription: Alex talking about advanced testing at Facebook London HQ
+---
